port(1.20.1): align investment correctness with 1.21.1 — config loop, treasury overflow, raid-force cap
The 1.20.1 line already wires all ten investments (getters are called), so the 1.21.1
"wire dormant investments" work is effectively already present here. What was missing were
the correctness fixes/hardenings from the 1.21.1 commits — three real defects present on
1.20.1 too:

- RaidTaxPercentages config self-correction loop (TaxConfig): it was the only list config
  defined with the scalar ForgeConfigSpec.define(path, value) overload, whose
  defaultValue.getClass().isAssignableFrom(loaded) check fails every load (ImmutableCollections
  ListN default vs. ArrayList parsed from TOML) — an infinite ~2 Hz rewrite/reload loop that
  floods the log. Switched to defineList(..., obj -> obj instanceof Double), matching every
  other list config here (e.g. EntityRaidWhitelist); field widened to List<? extends Double>.
  Both read sites in RaidManager stay compatible.

- Treasury cap int overflow (ColonyUpgradeManager.getTreasuryCapBonus +
  TreasuryManager.getEffectiveMaxCapacity): base + level*flat was int math, so an admin flat
  near Integer.MAX_VALUE overflowed to a negative effective cap (deposits rejected, headroom
  math broken). getTreasuryCapBonus now returns long and multiplies in long; getEffectiveMaxCapacity
  sums in long and clamps to [0, Integer.MAX_VALUE] so the cap only ever grows. All int callers
  unchanged.

- Raid Force over-steal (RaidManager.transferTaxRevenue): the RAID_FORCE multiplier scaled
  amountToDeduct with no cap and for BOTH the positive-balance and debt branches. A boosted
  percentage could steal more than the colony holds (creating debt from nothing) and could push
  the debt branch past the configured debt limit. Now scales only the positive-balance case and
  caps at colonyBalance — matching the 1.21.1 design ("debt branch left unscaled").

Deliberately NOT changed (flagged for the user): DEFENSE still requires base guard-resistance
to be enabled on 1.20.1 (the early-return remains), whereas 1.21.1's 50397e0 made DEFENSE apply
regardless. That is a behavioural/balance change, not a bug, so it is left as an open parity
decision rather than silently altered.

Build green on JDK 17 (Forge 1.20.1): compileJava, checkBuildingApiUsage, reobfJar, JUnit.

+++ src/main/java/net/machiavelli/minecolonytax/upgrade/ColonyUpgradeManager.java
@@ -123,8 +123,14 @@ public class ColonyUpgradeManager {
         return getLevel(colonyId, UpgradeType.DEFENSE) * TaxConfig.getUpgradeDefenseBonusPerLevel();
     }
 
-    public static int getTreasuryCapBonus(int colonyId) {
-        return getLevel(colonyId, UpgradeType.TREASURY_CAP) * TaxConfig.getUpgradeTreasuryCapFlatPerLevel();
+    /**
+     * Flat treasury capacity bonus from the TREASURY_CAP investment. Returns {@code long}: the
+     * per-level flat amount is admin-configurable up to Integer.MAX_VALUE, so {@code level * flat}
+     * can exceed the int range. Computing in long keeps the value exact for the single caller
+     * (TreasuryManager.getEffectiveMaxCapacity), which does its own [0, Integer.MAX_VALUE] clamp.
+     */
+    public static long getTreasuryCapBonus(int colonyId) {
+        return (long) getLevel(colonyId, UpgradeType.TREASURY_CAP) * TaxConfig.getUpgradeTreasuryCapFlatPerLevel();
     }
 
     public static double getTaxEfficiencyBonus(int colonyId) {
